requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
httpx==0.26.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic==2.10.6

# LLM / Agents (final compatible set)
langchain==0.2.3
langchain-community==0.2.4
langchain-groq==0.1.10
langgraph==0.0.69

# Embeddings & Vector DB
sentence-transformers==2.4.0
faiss-cpu==1.9.0.post1
numpy==1.26.4

# Document Parsing
PyMuPDF==1.25.4
python-docx==1.1.0

# Frontend
streamlit==1.31.1
plotly==5.18.0
networkx==3.2.1
websockets==12.0
requests==2.31.0